LPS-105380 Rename classes: Upgrade* -> *UpgradeProcess
