+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/upgrade/v1_0_1/RoleUpgradeProcess.java
@@ -29,7 +29,7 @@ import java.sql.ResultSet;
 /**
  * @author Pei-Jung Lan
  */
-public class UpgradeRole extends UpgradeProcess {
+public class RoleUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
