+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/upgrade/v1_0_2/RoleUpgradeProcess.java
@@ -22,7 +22,7 @@ import java.sql.PreparedStatement;
 /**
  * @author Pei-Jung Lan
  */
-public class UpgradeRole extends UpgradeProcess {
+public class RoleUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
