+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/upgrade/v1_1_0/SchemaUpgradeProcess.java
@@ -12,34 +12,23 @@
  * details.
  */
 
-package com.liferay.account.internal.upgrade.v1_2_1;
+package com.liferay.account.internal.upgrade.v1_1_0;
 
-import com.liferay.account.role.AccountRole;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.model.role.RoleConstants;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.StringUtil;
 
 /**
- * @author Drew Brokke
+ * @author Pei-Jung Lan
  */
-public class UpgradeRole extends UpgradeProcess {
+public class SchemaUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		int newType = RoleConstants.TYPE_ACCOUNT;
-		int oldType = RoleConstants.TYPE_PROVIDER;
+		String template = StringUtil.read(
+			SchemaUpgradeProcess.class.getResourceAsStream(
+				"dependencies/update.sql"));
 
-		String className = StringUtil.quote(
-			AccountRole.class.getName(), StringPool.APOSTROPHE);
-
-		runSQL(
-			StringBundler.concat(
-				"update Role_ set type_ = ", newType,
-				" where classNameId = (select ClassName_.classNameId from ",
-				"ClassName_ where ClassName_.value = ", className,
-				") and type_ = ", oldType));
+		runSQLTemplateString(template, false);
 	}
 
 }
