+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/upgrade/v1_2_1/RoleUpgradeProcess.java
@@ -12,29 +12,34 @@
  * details.
  */
 
-package com.liferay.account.internal.upgrade.v1_1_1;
+package com.liferay.account.internal.upgrade.v1_2_1;
 
-import com.liferay.account.constants.AccountConstants;
+import com.liferay.account.role.AccountRole;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.model.role.RoleConstants;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.StringUtil;
 
 /**
  * @author Drew Brokke
  */
-public class UpgradeAccountEntry extends UpgradeProcess {
+public class RoleUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		String oldType = StringUtil.quote("personal", StringPool.APOSTROPHE);
-		String newType = StringUtil.quote(
-			AccountConstants.ACCOUNT_ENTRY_TYPE_PERSON, StringPool.APOSTROPHE);
+		int newType = RoleConstants.TYPE_ACCOUNT;
+		int oldType = RoleConstants.TYPE_PROVIDER;
+
+		String className = StringUtil.quote(
+			AccountRole.class.getName(), StringPool.APOSTROPHE);
 
 		runSQL(
 			StringBundler.concat(
-				"update AccountEntry set type_ = ", newType, " where type_ = ",
-				oldType));
+				"update Role_ set type_ = ", newType,
+				" where classNameId = (select ClassName_.classNameId from ",
+				"ClassName_ where ClassName_.value = ", className,
+				") and type_ = ", oldType));
 	}
 
 }
