+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/upgrade/v1_3_0/AccountEntryUpgradeProcess.java
@@ -14,20 +14,39 @@
 
 package com.liferay.account.internal.upgrade.v1_3_0;
 
-import com.liferay.account.internal.upgrade.v1_3_0.util.AccountGroupTable;
+import com.liferay.account.internal.upgrade.v1_3_0.util.AccountEntryTable;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 
 /**
  * @author Pei-Jung Lan
  */
-public class UpgradeAccountGroup extends UpgradeProcess {
+public class AccountEntryUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		if (!hasColumn("AccountGroup", "defaultAccountGroup")) {
+		if (!hasColumn("AccountEntry", "defaultBillingAddressId")) {
 			alter(
-				AccountGroupTable.class,
-				new AlterTableAddColumn("defaultAccountGroup", "BOOLEAN"));
+				AccountEntryTable.class,
+				new AlterTableAddColumn("defaultBillingAddressId", "LONG"));
+		}
+
+		if (!hasColumn("AccountEntry", "defaultShippingAddressId")) {
+			alter(
+				AccountEntryTable.class,
+				new AlterTableAddColumn("defaultShippingAddressId", "LONG"));
+		}
+
+		if (!hasColumn("AccountEntry", "emailAddress")) {
+			alter(
+				AccountEntryTable.class,
+				new AlterTableAddColumn("emailAddress", "VARCHAR(254) null"));
+		}
+
+		if (!hasColumn("AccountEntry", "taxExemptionCode")) {
+			alter(
+				AccountEntryTable.class,
+				new AlterTableAddColumn(
+					"taxExemptionCode", "VARCHAR(75) null"));
 		}
 	}
